Limited playHistory to below 11 :bug:
Note: We might want `playHistory` to keep indexes of all played tracks in shuffle so the user has the impression of persisted play history.

The issue is what if the user switches to a different context (i.e. a different playlist, all Tracks, a different album, etc) and two separate tracks have the same index, we can't then assume the new track has already been played.