player: EQ gain is split into alpha & beta channel
- We have split the player's gain value to two: the EQ preamp (alpha channel) and the replayGain value (beta channel).

- The two gains are therefore normalized at a limit of 0.5 each since the gain value is between 0 - 1 inclusive.

- Added a dedicated fn to return preamp's normalized value.

- Separated preamp value setting fn, it seems sometimes changing the values of the other channels affects the preamp.

Note: the preamp is not exactly a channel that controls a specific frequency, instead it acts as a gain modulator for all the frequencies.

Attention: We still need to run more tests on our replayGain implementation, especially since the gain is now a combined value from both the calculated replayGain and EQ preamp.

Caution: The gain value seems to be a bit unpredictable even when switching in/out of the a specific EQ preset.